tidy rpc module and move it to `utils/`
