+++ randchain/utils/mod.rs
@@ -1,4 +1,5 @@
 mod db;
+pub mod rpc;
 mod seednodes;
 mod sync;
